add code to discounts and cart_items to delete simple change cart view to show multiple discounts

+++ Modules/Shop/database/migrations/2025_08_14_122755_create_discounts_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('code')->nullable(); // for manual code entry
             $table->enum('type', ['percent', 'fixed']);
-            $table->decimal('value', 10, 2);
+            $table->bigInteger('value');
             $table->timestamp('start_at')->nullable();
             $table->timestamp('end_at')->nullable();
             $table->boolean('is_active')->default(true);
